+++ mobileapp/app/src/main/java/com/example/mobileapp/features/shared/api/dto/RideDto.java
@@ -1,0 +1,13 @@
+package com.example.mobileapp.features.shared.api.dto;
+
+public class RideDto {
+    public Integer id;
+    public String date;
+    public String time;
+    public String from;
+    public String to;
+    public String status;
+    public String cancelledBy;
+    public Boolean panic;
+    public String price;
+}
