+++ mobileapp/app/src/main/java/com/example/mobileapp/features/shared/api/dto/RideHistoryResponseDto.java
@@ -1,0 +1,10 @@
+package com.example.mobileapp.features.shared.api.dto;
+
+import java.util.List;
+
+public class RideHistoryResponseDto {
+    public List<RideDto> rides;
+    public long total;
+    public int page;
+    public int size;
+}
