+++ mobileapp/app/src/main/java/com/example/mobileapp/features/shared/models/VehicleMarker.java
@@ -1,0 +1,10 @@
+package com.example.mobileapp.features.shared.models;
+
+import com.example.mobileapp.features.shared.models.enums.VehicleStatus;
+
+public class VehicleMarker {
+    public int id;
+    public double lat;
+    public double lng;
+    public VehicleStatus status;
+}
