+++ mobileapp/app/src/main/java/com/example/mobileapp/features/shared/models/enums/RideStatus.java
@@ -1,0 +1,8 @@
+package com.example.mobileapp.features.shared.models.enums;
+
+public enum RideStatus {
+    COMPLETED,
+    CANCELLED,
+    ASSIGNED,
+    SCHEDULED
+}
