+++ mobileapp/app/src/main/java/com/example/mobileapp/features/shared/models/enums/VehicleStatus.java
@@ -1,0 +1,7 @@
+package com.example.mobileapp.features.shared.models.enums;
+
+public enum VehicleStatus {
+    AVAILABLE,
+    OCCUPIED,
+    INACTIVE
+}
