--- mobileapp/app/src/main/java/com/example/mobileapp/model/RideStatus.java
@@ -1,8 +1,0 @@
-package com.example.mobileapp.model;
-
-public enum RideStatus {
-    COMPLETED,
-    CANCELLED,
-    ASSIGNED,
-    SCHEDULED
-}
