--- mobileapp/app/src/main/java/com/example/mobileapp/model/VehicleMarker.java
@@ -1,8 +1,0 @@
-package com.example.mobileapp.model;
-
-public class VehicleMarker {
-    public int id;
-    public double lat;
-    public double lng;
-    public VehicleStatus status;
-}
