--- mobileapp/app/src/main/java/com/example/mobileapp/model/VehicleStatus.java
@@ -1,7 +1,0 @@
-package com.example.mobileapp.model;
-
-public enum VehicleStatus {
-    AVAILABLE,
-    OCCUPIED,
-    INACTIVE
-}
