--- mobileapp/app/src/main/java/com/example/mobileapp/network/ApiClient.java
@@ -1,24 +1,0 @@
-package com.example.mobileapp.network;
-
-import retrofit2.Retrofit;
-import retrofit2.converter.gson.GsonConverterFactory;
-
-public final class ApiClient {
-
-    private static final String BASE_URL = "http://192.168.50.49:8080/";
-    // private static final String BASE_URL = "http://10.0.2.2:8080/";
-
-    private static Retrofit retrofit;
-
-    private ApiClient() {}
-
-    public static Retrofit get() {
-        if (retrofit == null) {
-            retrofit = new Retrofit.Builder()
-                    .baseUrl(BASE_URL)
-                    .addConverterFactory(GsonConverterFactory.create())
-                    .build();
-        }
-        return retrofit;
-    }
-}
