--- mobileapp/app/src/main/java/com/example/mobileapp/network/dto/RideDto.java
@@ -1,13 +1,0 @@
-package com.example.mobileapp.network.dto;
-
-public class RideDto {
-    public Integer id;
-    public String date;
-    public String time;
-    public String from;
-    public String to;
-    public String status;
-    public String cancelledBy;
-    public Boolean panic;
-    public String price;
-}
