--- mobileapp/app/src/main/java/com/example/mobileapp/network/dto/RideHistoryResponseDto.java
@@ -1,10 +1,0 @@
-package com.example.mobileapp.network.dto;
-
-import java.util.List;
-
-public class RideHistoryResponseDto {
-    public List<RideDto> rides;
-    public long total;
-    public int page;
-    public int size;
-}
