--- mobileapp/app/src/main/java/com/example/mobileapp/ui/LoginFragment.java
@@ -1,54 +1,0 @@
-package com.example.mobileapp.ui;
-
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.Button;
-import android.widget.EditText;
-import android.widget.TextView;
-
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import androidx.fragment.app.Fragment;
-
-import com.example.mobileapp.R;
-
-public class LoginFragment extends Fragment {
-
-    private EditText etEmail, etPassword;
-    private Button btnLogin;
-    private TextView tvForgotPassword, tvSignUp;
-
-    @Nullable
-    @Override
-    public View onCreateView(@NonNull LayoutInflater inflater,
-                             @Nullable ViewGroup container,
-                             @Nullable Bundle savedInstanceState) {
-        // Use your existing layout
-        View view = inflater.inflate(R.layout.fragment_login, container, false);
-
-        initViews(view);
-        setupClickListeners();
-
-        return view;
-    }
-
-    private void initViews(View view) {
-        etEmail = view.findViewById(R.id.etEmail);
-        etPassword = view.findViewById(R.id.etPassword);
-        btnLogin = view.findViewById(R.id.btnLogin);
-        tvForgotPassword = view.findViewById(R.id.tvForgotPassword);
-        tvSignUp = view.findViewById(R.id.tvSignUp);
-    }
-
-    private void setupClickListeners() {
-        tvSignUp.setOnClickListener(v -> {
-            requireActivity().getSupportFragmentManager()
-                    .beginTransaction()
-                    .replace(R.id.fragment_container_view, new RegisterFragment())
-                    .addToBackStack(null)
-                    .commit();
-        });
-    }
-}
