+++ mobileapp/app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".ui.MainActivity">
+    tools:context=".DriverMainActivity">
 
     <!-- Main screen content (toolbar + fragment area) -->
     <LinearLayout
