+++ mobileapp/app/src/main/res/layout/activity_user_main.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".ui.UserMainActivity">
+    tools:context=".UserMainActivity">
 
     <!-- Main screen content (toolbar + fragment area) -->
     <LinearLayout
