+++ mobileapp/app/src/main/res/layout/dialog_fragment_change_password.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:padding="20dp"
-    tools:context=".ui.ChangePasswordDialogFragment">
+    tools:context=".features.driver.profile.ChangePasswordDialogFragment">
 
     <androidx.cardview.widget.CardView
         android:layout_width="match_parent"
