+++ mobileapp/app/src/main/res/layout/fragment_profile.xml
@@ -8,7 +8,7 @@
     android:fillViewport="true"
     android:background="@color/app_bg"
     android:paddingBottom="24dp"
-    tools:context=".ui.ProfileFragment">
+    tools:context=".features.driver.profile.ProfileFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
